ldelf: Add CFG_TA_PIE to help building TAs for non-PIE capable targets
This commit introduces CFG_TA_PIE flag to compile Trusted Applications
as position-independent executables or not.
The idea, is to be able to build OP-TEE for targets which are not ASLR
capable.
If enabled, it adds -fpie flag to compile TA code and link using -pie
option. Enabling CFG_TA_ASLR depends on CFG_TA_PIE being enabled.
On arm target, user-space must be relocatable, hence, CFG_TA_PIE is enabled
unconditionally.

Signed-off-by: Marouene Boubakri <marouene.boubakri@nxp.com>