core: simple framework to execute deferred works
This patch adds to the kernel a new framework for the ability
to save different works, when Rich OS is still booting
and tee-supplicant has not started yet.

Also the patch adds small PTA to trigger different works for execution.
tee-supplicant starts and sends a cmd to the PTA to run all pending jobs.

This ability might be useful in different cases. For example:
 - Run some specific RPC as soon as possible, when RichOS will start;
 - Remove REE based key store or any other fs operations;
 - Save TEE boot process log

To create a new deferred work entity - use 'deferred_work_add()'.
The works will collect in a queue in the kernel.

Signed-off-by: Aleksandr Anisimov <a.anisimov@omprussia.ru>