plat-ls: register dynamic shared memory for lx2160a
With the recent updates in U-Boot and OP-TEE this platform
can store it's EFI variables in the RPMB partition of it's eMMC
device.
In order for that to happen U-Boot must be able to access OP-TEE.
Since U-Boot only implements dynamic shared memory communication at
the moment, let's register the whole DRAM minus the memory
used for Trusted RAM and the untrusted static shared memory.

Signed-off-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>