scripts : Check the input already sign before signing a TA
The TA will be re-signed, if the sign script is called multiple times.
Just like the script below was executed twice and the final TA will be
Signature twice.

python scripts/sign_encrypt.py sign-enc --key ./keys/default_ta.pem
--uuid ${uuid} --ta-version 0 --in ${uuid}.ta --out ${uuid}.ta

So, at the time of signing, it should check whether the input TA
is signed. If signed , it will remove the header information and
signature and then be signed.

Signed-off-by: Baron Zhou <zhhh891010@163.com>