core: recreate a new hash file when dirfile is not found
When CFG_RPMB_FS is y, OP-TEE stores a hash file in RPMB to protect
the REE FS. If the REE RS is cleaned without cleaning RPMB, the
dirfile will be inconsisten with the hash file. That is,when one TA
calls TEE_OpenPersistentObject, is causes TA panic with
TEE_ERROR_SECURITY in this situation. Therefore, recreates a new hash
when the dirfile is not found.

Signed-off-by: Rayan Hu <Rayan.Hu@mediatek.com>
Signed-off-by: Menson Chen <Menson.Chen@mediatek.com>