prng: provide prng when hwrng is enabled
Secure Channel Protocol 03 requires that all communication between the
secure element and the processor is software encrypted.

This means that in order to read the RNG provided by the secure
elements, we first need to gather some entropy to feed to the
software cryptograpnhic algorithms.

This commit alters the meaning of CFG_WITH_SOFTWARE_PRNG:

1) When enabled, crypto_prng_* should not be used as crypto_rng_*
already provides the randomness.

2) When not enabled, crypto_rng_* provides the HWRNG functionality
while crypto_prng_* provides the PRNG (fortuna) randomness.

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>