core: scmi_server: Build a SCMI server from SCP-firmware source tree
Adds build of SCP-firmware to build platform SCMI services as a library
and integration in the SCMI PTA. Configuration switch CFG_SCMI_SERVER=y|n
defines in the server is embedded of not.

CFG_SCMI_SERVER and CFG_SCMI_MSG_DRIVERS are exclusives, that is either
SCMI server is built on scmi-msg drivers or the SCMI server is fully
provided by the embedded SCP-firmware library.

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>