core: scmi_server: Build a SCMI server from SCP-firmware
Add the SCP-firmware as a SCMI server to be integrated in the SCMI PTA.
The configuration switch CFG_SCMI_SERVER=y selects the SCP-firmware to
be used to implement SCMI services.

CFG_SCMI_SERVER and CFG_SCMI_MSG_DRIVERS are exclusives and the SCMI
server is built on either scmi-msg drivers or the SCMI server.

SCP-firmware is built with CMake as an external project. The source path
is defined with CFG_SCP_FIRMWARE.

Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>