core: FF-A: ADD FFA_MEM_SHARE for SPs
When receiving a FFA_MEM_SHARE FF-A message:
The whole mem share buffer (mem_transaction_descr) is stored. A
shared_mem object is created. The struct shared_mem is used for
bookkeeping and access to the struct mem_transaction_descr.
It also contains a list sp_shared_mem objects. These create the link
between the endpoint_id and the ffa_mem_access struct(SP specific data
in a memory share).

Signed-off-by: Jelle Sels <jelle.sels@arm.com>