core: ldelf: create .debug_frame when using AARCH64 GCC
This patch prevents the generation of the .eh_frame using GCC command
line options. By the same time GCC starts generating .debug_frame
instead. This improves the robustness for call-stack unwinding when
using an external debug probe.
As .eh_frame was removed in the linker scripts this is no functional
change.

Signed-off-by: Alexander Merkle <alexander.merkle@lauterbach.com>