tee_rpmb_fs: Adding offset to memref to ensure 32-bit aligment
Modifying thread_param struct to add the correct offset.
The `struct rpmb_req` with the offset reaches a size modulo 32bits.
As the memory allocation size is already roundup to 32bits, no need
to add the offset as it is calculated to reach a roundup of 32bits.

The RPMB data frame is located after `struct rpmb_req` in memory, its
address needs a 32 bits alignment to be passed to U-Boot RPMB driver.
Adding offset to `struct rpmb_req` fixes this and doesn't break ABI.

Signed-off-by: Timothée Cercueil <timothee.cercueil@st.com>
Change-Id: Ic9b223363ada9044db00a0f30eb80f8898b75181