core: FF-A: ADD FFA_MEM_SHARE for SPs
Process a FF-A FFA_MEM_SHARE message coming from a SP or being send
from the Normal world with one or more SPs a receivers.
FFA_MEM_SHARE is used to share a memory region from an endpoint (SP or
normal world) with one or more endpoints in de secure world(SPs).

A simplified version of the share memory transaction descriptor looks
the following:

|-------------------|
|ffa_mem_transaction| -> Global information (Stored inside mobj_ffa)
|-------------------|
|mem_access_array[0]| -> Receiver specific information
|                   |    (Stored inside sp_mem_access_descr)
|-------------------|
|mem_access_array[1]|
|-------------------|
|mem_access_array[n]|
|-------------------|
|ffa_mem_region     |-> Global information (Data stored inside page
|                   |   are of mobj_ffa)
|-------------------|

-The descriptor has parts which are the same for each receiving
endpoint: ffa_mem_transaction and ffa_mem_region. These are the same
for every receiver.
For every memory share we create one mobj_ffa object. This is used to
control and keep track of the memory regions. The unique cookie is used
the FF-A memory share handle. And can be used to identify the memory
share.
A ffa_mem_transaction object is added inside the mobj_ffa.
The ffa_mem_region specific data is stored inside the pages area of the
struct mobj_ffa.

-The descriptor also contains some receiver specific information:
mem_access_array[]. Each element points to information that is specific
for a memory share SP receiver. This data is stored inside a struct
sp_mem_access_descr object.

Signed-off-by: Jelle Sels <jelle.sels@arm.com>